requirements.txt
langgraph
langchain
langchain-groq
langchain-community
duckduckgo-search
ddgs
python-dotenv
streamlit